Final Code
Added in the code for the project. Changed the files to their proper names. Added in code to select the rows and columns. Added in code to use the arrow keys and type in letters. Added in code to show the solution. Added in the code to check for errors.